Remove .env and add all the usual suspects to .gitignore.
The .env should only be copied from env.sample and
the .gitignore should help to prevent .env files from
accidentally being added to git commits.

The README and env.sample is already done. Just need to remove
the old one and setup the .gitignore.

The default .gitignore file for node projects was used to
make our .gitignore ignore typical things that we might see
in the future.